renaming and back auto + delete
