+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Autonomous/BFRightNOSplines.java
@@ -15,8 +15,8 @@ import org.firstinspires.ftc.teamcode.Variables.VisionProcessors;
 import java.util.Arrays;
 
 @Config
-@Autonomous(name = "RCInLeft (Testing Down)", group = "Linear OpMode")
-public class RCInLeft extends MeepMeepBoilerplate{
+@Autonomous(name = "BFRightNOSplines", group = "Linear OpMode")
+public class BFRightNOSplines extends MeepMeepBoilerplate{
     @Override
     public void runOpMode() {
         SampleMecanumDrive drive = new SampleMecanumDrive(hardwareMap);
@@ -24,9 +24,9 @@ public class RCInLeft extends MeepMeepBoilerplate{
         Servo autoServo = hardwareMap.get(Servo.class, "autoServo");
         initVision(VisionProcessors.TFOD);
         Detection detection = Detection.UNKNOWN;
-        TrajectoryVelocityConstraint fastConstraint = new MinVelocityConstraint(Arrays.asList(
+        TrajectoryVelocityConstraint slowConstraint = new MinVelocityConstraint(Arrays.asList(
 
-                new TranslationalVelocityConstraint(50),
+                new TranslationalVelocityConstraint(20),
 
                 new AngularVelocityConstraint(1)
 
@@ -36,11 +36,15 @@ public class RCInLeft extends MeepMeepBoilerplate{
             telemetry.addData("Detection", detection);
             telemetry.update();
         }
-        autoServo.setPosition(0.78);
         switch (detection) {
             case LEFT -> drive.followTrajectorySequence(
                     drive.trajectorySequenceBuilder(getCurrentPosition(drive))
-                            .addTemporalMarker(() -> goDownAuto())
+                            .back(28.0)
+                            .turn(Math.toRadians(90))
+                            .back(8)
+                            .addTemporalMarker(() -> passiveServo.setPosition(0.2))
+                            .waitSeconds(4)
+                            .forward(5)
                             .build()
             );
             case CENTER -> { drive.followTrajectorySequence(
@@ -48,25 +52,8 @@ public class RCInLeft extends MeepMeepBoilerplate{
                             .back(31.5)
                             .waitSeconds(.25)
                             .addTemporalMarker(() -> passiveServo.setPosition(0.2))
-                            .waitSeconds(.25)
-                            .forward(10)
-                            .waitSeconds(.25)
-                            .turn(Math.toRadians(-90))
-                            .waitSeconds(.25)
-                            .back(12)
-                            .strafeRight(9)
-                            .back(24.5)
-                            .waitSeconds(.25)
-                            .addTemporalMarker(() -> autoServo.setPosition(1))
-                            .waitSeconds(.5)
-                            .addTemporalMarker(() -> autoServo.setPosition(0.9))
-                            .waitSeconds(.25)
-                            .forward(3)
-                            .strafeRight(20)
-                            .back(11)
-                            .waitSeconds(1)
-                            .addTemporalMarker(() -> autoServo.setPosition(0.65))
-                            .waitSeconds(1)
+                            .waitSeconds(4)
+                            .forward(5)
                             .build());
             }
             case RIGHT -> drive.followTrajectorySequence(
@@ -77,27 +64,8 @@ public class RCInLeft extends MeepMeepBoilerplate{
                             .back(21)
                             .waitSeconds(.25)
                             .addTemporalMarker(() -> passiveServo.setPosition(0.2))
-                            .waitSeconds(.25)
-                            .forward(9.5)
-                            .waitSeconds(.25)
-                            .turn(Math.toRadians(-90))
-                            .waitSeconds(.25)
-                            .back(12.5)
-                            .waitSeconds(.25)
-                            .strafeRight(12)
-                            .waitSeconds(.25)
-                            .back(15.5)
-                            .waitSeconds(.25)
-                            .addTemporalMarker(() -> autoServo.setPosition(1))
-                            .waitSeconds(.5)
-                            .addTemporalMarker(() -> autoServo.setPosition(0.9))
-                            .waitSeconds(.25)
-                            .forward(3)
-                            .strafeRight(26)
-                            .back(10)
-                            .waitSeconds(1)
-                            .addTemporalMarker(() -> autoServo.setPosition(0.65))
-                            .waitSeconds(1)
+                            .waitSeconds(4)
+                            .forward(5)
                             .build()
             );
             default -> {
@@ -106,7 +74,6 @@ public class RCInLeft extends MeepMeepBoilerplate{
                 sleep(3000);
             }
         }
-
 
 
 //        drive.followTrajectorySequence(mergeSequences(sequences));
