--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Autonomous/RFRight.java
@@ -1,82 +1,0 @@
-package org.firstinspires.ftc.teamcode.Autonomous;
-
-import com.acmerobotics.dashboard.config.Config;
-import com.acmerobotics.roadrunner.trajectory.constraints.AngularVelocityConstraint;
-import com.acmerobotics.roadrunner.trajectory.constraints.MinVelocityConstraint;
-import com.acmerobotics.roadrunner.trajectory.constraints.TrajectoryVelocityConstraint;
-import com.acmerobotics.roadrunner.trajectory.constraints.TranslationalVelocityConstraint;
-import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
-import com.qualcomm.robotcore.hardware.Servo;
-
-import org.firstinspires.ftc.teamcode.RoadRunner.drive.SampleMecanumDrive;
-import org.firstinspires.ftc.teamcode.Variables.Detection;
-import org.firstinspires.ftc.teamcode.Variables.VisionProcessors;
-
-import java.util.Arrays;
-
-@Config
-@Autonomous(name = "RFRight(Actual)", group = "LinearOpmode")
-public class RFRight extends MeepMeepBoilerplate {
-
-    @Override
-    public void runOpMode() {
-        SampleMecanumDrive drive = new SampleMecanumDrive(hardwareMap);
-        Servo passiveServo = hardwareMap.get(Servo.class, "passiveServo");
-        Servo autoServo = hardwareMap.get(Servo.class, "autoServo");
-        initVision(VisionProcessors.TFOD);
-        Detection detection = Detection.UNKNOWN;
-        TrajectoryVelocityConstraint slowConstraint = new MinVelocityConstraint(Arrays.asList(
-
-                new TranslationalVelocityConstraint(20),
-
-                new AngularVelocityConstraint(1)
-
-        ));
-        while (opModeInInit()) {
-            detection = getDetectionsSingleTFOD();
-            telemetry.addData("Detection", detection);
-            telemetry.update();
-        }
-
-        switch (detection) {
-            case LEFT -> drive.followTrajectorySequence(
-                    drive.trajectorySequenceBuilder(getCurrentPosition(drive))
-                            .back(2.0)
-                            .waitSeconds(.1)
-                            .strafeRight(11)
-                            .back(21)
-                            .waitSeconds(.1)
-                            .addTemporalMarker(() -> passiveServo.setPosition(0.2))
-                            .waitSeconds(4)
-                            .forward(5)
-                            .build()
-            );
-            case CENTER -> { drive.followTrajectorySequence(
-                    drive.trajectorySequenceBuilder(getCurrentPosition(drive))
-                            .back(32.25)
-                            .waitSeconds(.25)
-                            .addTemporalMarker(() -> passiveServo.setPosition(0.2))
-                            .waitSeconds(4)
-                            .forward(5)
-                            .build());
-            }
-            case RIGHT -> drive.followTrajectorySequence(
-                    drive.trajectorySequenceBuilder(getCurrentPosition(drive))
-                            .back(28.0)
-                            .turn(Math.toRadians(-90))
-                            .back(4)
-                            .addTemporalMarker(() -> passiveServo.setPosition(0.2))
-                            .waitSeconds(4)
-                            .forward(5)
-                            .build()
-            );
-            default -> {
-                telemetry.addLine("Warning: Cup not detected");
-                telemetry.update();
-                sleep(3000);
-            }
-        }
-
-//        drive.followTrajectorySequence(mergeSequences(sequences));
-    }
-}
